atari:video:gem: Atari_get_EdDI_version() returns 16-bit int

+++ src/video/ataricommon/SDL_atarieddi_s.h
@@ -49,6 +49,6 @@ enum {
 
 /*--- Functions ---*/
 
-unsigned long Atari_get_EdDI_version(void *function_pointer);
+unsigned short Atari_get_EdDI_version(void *function_pointer);
 
 #endif /* _SDL_Atari_eddi_s_h */
